Rename test classes to remove unnecessary Bespoke name
