+++ java/ingest/ingest-runner/src/test/java/sleeper/ingest/impl/IngestCoordinatorUsingDirectWriteBackedByArrowIT.java
@@ -15,6 +15,7 @@
  */
 package sleeper.ingest.impl;
 
+import org.apache.arrow.memory.OutOfMemoryException;
 import org.apache.commons.lang3.tuple.Pair;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
@@ -28,7 +29,7 @@ import sleeper.core.key.Key;
 import sleeper.core.record.Record;
 import sleeper.core.schema.type.LongType;
 import sleeper.ingest.impl.partitionfilewriter.DirectPartitionFileWriterFactory;
-import sleeper.ingest.impl.recordbatch.arraylist.ArrayListRecordBatchFactory;
+import sleeper.ingest.impl.recordbatch.arrow.ArrowRecordBatchFactory;
 import sleeper.ingest.testutils.AwsExternalResource;
 import sleeper.ingest.testutils.PartitionedTableCreator;
 import sleeper.ingest.testutils.RecordGenerator;
@@ -48,10 +49,11 @@ import java.util.stream.LongStream;
 import java.util.stream.Stream;
 
 import static java.nio.file.Files.createTempDirectory;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static sleeper.ingest.testutils.IngestCoordinatorTestHelper.parquetConfiguration;
 import static sleeper.ingest.testutils.IngestCoordinatorTestHelper.standardIngestCoordinator;
 
-public class IngestCoordinatorBespokeUsingDirectWriteBackedByArrayListIT {
+public class IngestCoordinatorUsingDirectWriteBackedByArrowIT {
     @RegisterExtension
     public static final AwsExternalResource AWS_EXTERNAL_RESOURCE = new AwsExternalResource(
             LocalStackContainer.Service.S3,
@@ -74,7 +76,7 @@ public class IngestCoordinatorBespokeUsingDirectWriteBackedByArrayListIT {
     public void shouldWriteRecordsWhenThereAreMoreRecordsInAPartitionThanCanFitInMemory() throws Exception {
         RecordGenerator.RecordListAndSchema recordListAndSchema = RecordGenerator.genericKey1D(
                 new LongType(),
-                LongStream.range(-100, 100).boxed().collect(Collectors.toList()));
+                LongStream.range(-10000, 10000).boxed().collect(Collectors.toList()));
         List<Pair<Key, Integer>> keyAndDimensionToSplitOnInOrder = Collections.singletonList(
                 Pair.of(Key.create(0L), 0));
         Function<Key, Integer> keyToPartitionNoMappingFn = key -> (((Long) key.get(0)) < 0L) ? 0 : 1;
@@ -82,43 +84,71 @@ public class IngestCoordinatorBespokeUsingDirectWriteBackedByArrayListIT {
                         new AbstractMap.SimpleEntry<>(0, 1),
                         new AbstractMap.SimpleEntry<>(1, 1))
                 .collect(Collectors.toMap(Map.Entry::getKey, Map.Entry::getValue));
-        ingestAndVerifyUsingDirectWriteBackedByArrayList(
+        ingestAndVerifyUsingDirectWriteBackedByArrow(
                 recordListAndSchema,
                 keyAndDimensionToSplitOnInOrder,
                 keyToPartitionNoMappingFn,
                 partitionNoToExpectedNoOfFilesMap,
-                5,
-                1000L);
+                16 * 1024 * 1024L,
+                4 * 1024 * 1024L,
+                128 * 1024 * 1024L);
     }
 
     @Test
     public void shouldWriteRecordsWhenThereAreMoreRecordsThanCanFitInLocalFile() throws Exception {
         RecordGenerator.RecordListAndSchema recordListAndSchema = RecordGenerator.genericKey1D(
                 new LongType(),
-                LongStream.range(-100, 100).boxed().collect(Collectors.toList()));
+                LongStream.range(-10000, 10000).boxed().collect(Collectors.toList()));
         List<Pair<Key, Integer>> keyAndDimensionToSplitOnInOrder = Collections.singletonList(
                 Pair.of(Key.create(0L), 0));
         Function<Key, Integer> keyToPartitionNoMappingFn = key -> (((Long) key.get(0)) < 0L) ? 0 : 1;
         Map<Integer, Integer> partitionNoToExpectedNoOfFilesMap = Stream.of(
-                        new AbstractMap.SimpleEntry<>(0, 20),
-                        new AbstractMap.SimpleEntry<>(1, 20))
+                        new AbstractMap.SimpleEntry<>(0, 2),
+                        new AbstractMap.SimpleEntry<>(1, 2))
                 .collect(Collectors.toMap(Map.Entry::getKey, Map.Entry::getValue));
-        ingestAndVerifyUsingDirectWriteBackedByArrayList(
+        ingestAndVerifyUsingDirectWriteBackedByArrow(
                 recordListAndSchema,
                 keyAndDimensionToSplitOnInOrder,
                 keyToPartitionNoMappingFn,
                 partitionNoToExpectedNoOfFilesMap,
-                5,
-                10L);
+                16 * 1024 * 1024L,
+                4 * 1024 * 1024L,
+                16 * 1024 * 1024L);
     }
 
-    private void ingestAndVerifyUsingDirectWriteBackedByArrayList(
+    @Test
+    public void shouldErrorWhenBatchBufferAndWorkingBufferAreSmall() {
+        RecordGenerator.RecordListAndSchema recordListAndSchema = RecordGenerator.genericKey1D(
+                new LongType(),
+                LongStream.range(-10000, 10000).boxed().collect(Collectors.toList()));
+        List<Pair<Key, Integer>> keyAndDimensionToSplitOnInOrder = Collections.singletonList(
+                Pair.of(Key.create(0L), 0));
+        Function<Key, Integer> keyToPartitionNoMappingFn = key -> (((Long) key.get(0)) < 0L) ? 0 : 1;
+        Map<Integer, Integer> partitionNoToExpectedNoOfFilesMap = Stream.of(
+                        new AbstractMap.SimpleEntry<>(0, 2),
+                        new AbstractMap.SimpleEntry<>(1, 2))
+                .collect(Collectors.toMap(Map.Entry::getKey, Map.Entry::getValue));
+        assertThatThrownBy(() ->
+                ingestAndVerifyUsingDirectWriteBackedByArrow(
+                        recordListAndSchema,
+                        keyAndDimensionToSplitOnInOrder,
+                        keyToPartitionNoMappingFn,
+                        partitionNoToExpectedNoOfFilesMap,
+                        32 * 1024L,
+                        1024 * 1024L,
+                        64 * 1024 * 1024L))
+                .isInstanceOf(OutOfMemoryException.class)
+                .hasNoSuppressedExceptions();
+    }
+
+    private void ingestAndVerifyUsingDirectWriteBackedByArrow(
             RecordGenerator.RecordListAndSchema recordListAndSchema,
             List<Pair<Key, Integer>> keyAndDimensionToSplitOnInOrder,
             Function<Key, Integer> keyToPartitionNoMappingFn,
             Map<Integer, Integer> partitionNoToExpectedNoOfFilesMap,
-            int maxNoOfRecordsInMemory,
-            long maxNoOfRecordsInLocalStore) throws IOException, StateStoreException, IteratorException {
+            long arrowWorkingBytes,
+            long arrowBatchBytes,
+            long localStoreBytes) throws IOException, StateStoreException, IteratorException {
         StateStore stateStore = PartitionedTableCreator.createStateStore(
                 AWS_EXTERNAL_RESOURCE.getDynamoDBClient(),
                 recordListAndSchema.sleeperSchema,
@@ -129,11 +159,14 @@ public class IngestCoordinatorBespokeUsingDirectWriteBackedByArrayListIT {
                 recordListAndSchema.sleeperSchema, AWS_EXTERNAL_RESOURCE.getHadoopConfiguration());
         try (IngestCoordinator<Record> ingestCoordinator = standardIngestCoordinator(
                 stateStore, recordListAndSchema.sleeperSchema,
-                ArrayListRecordBatchFactory.builder()
-                        .parquetConfiguration(parquetConfiguration)
+                ArrowRecordBatchFactory.builder()
+                        .schema(recordListAndSchema.sleeperSchema)
+                        .maxNoOfRecordsToWriteToArrowFileAtOnce(128)
+                        .workingBufferAllocatorBytes(arrowWorkingBytes)
+                        .minBatchBufferAllocatorBytes(arrowBatchBytes)
+                        .maxBatchBufferAllocatorBytes(arrowBatchBytes)
+                        .maxNoOfBytesToWriteLocally(localStoreBytes)
                         .localWorkingDirectory(ingestLocalWorkingDirectory)
-                        .maxNoOfRecordsInMemory(maxNoOfRecordsInMemory)
-                        .maxNoOfRecordsInLocalStore(maxNoOfRecordsInLocalStore)
                         .buildAcceptingRecords(),
                 DirectPartitionFileWriterFactory.from(
                         parquetConfiguration, "s3a://" + DATA_BUCKET_NAME))) {
